@@ -4,17 +4,13 @@
   "private": true,
   "homepage": "https://massikv.github.io/bureaux/",
   "dependencies": {
-    "@auth0/auth0-react": "^2.2.1",
     "@reduxjs/toolkit": "^1.9.2",
-    "antd": "^5.0.7",
-    "bootstrap": "^5.3.1",
-    "bootstrap-scss": "^5.2.3",
     "dayjs": "^1.11.7",
-    "react": "^16.7.0",
+    "gsap": "^3.12.2",
+    "lodash": "^4.17.21",
+    "react": "^18.2.0",
     "react-alice-carousel": "^2.7.1",
-    "react-bootstrap": "^2.8.0",
-    "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^16.7.0",
+    "react-dom": "^18.2.0",
     "react-form-stepper": "^2.0.3",
     "react-modal": "^3.16.1",
     "react-number-format": "^5.1.1",
@@ -22,13 +18,11 @@
     "react-phone-input-2": "^2.15.1",
     "react-redux": "^8.0.5",
     "react-responsive-carousel": "^3.2.23",
-    "react-reveal": "^1.2.2",
     "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
     "react-scroll": "^1.8.9",
-    "react-scroll-horizontal": "^1.6.6",
+    "react-scroll-trigger": "^0.6.14",
     "react-spring": "^9.7.2",
-    "react-ticker": "1.3.0",
     "sass": "^1.57.0",
     "web-vitals": "^2.1.4"
   },
@@ -46,7 +40,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "deploy":"gh-pages -d build" ,
+    "deploy": "gh-pages -d build",
     "predeploy": "npm run build"
   },
   "eslintConfig": {
